Split combat text into separate header
