+++ FirstPersonScience.vcxproj
@@ -132,6 +132,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="source\App.h" />
+    <ClInclude Include="source\CombatText.h" />
     <ClInclude Include="source\Session.h" />
     <ClInclude Include="source\ConfigFiles.h" />
     <ClInclude Include="source\Logger.h" />
